Fixing the bug with no openstreetmap.org geometry for that _location/radius_/shapeType_ inputs
The following bug has been fixed:
If there is no openstreetmap.org geometry for the chosen "geometryType_" input (example: you have geometryType_ = 2 in a place where there are no points on openstreetmap.org), up until now the "OSM Shapes" raised a warning that one needs to disconnect the values from the ```onlyRemove_Ids_``` input even when this input was empty. This has now been fixed, and a new warning is raised informing the user that on that _location/radius_/shapeType_ no openstreetmap.org geometry exists.