Catching the "OSM 3D" error message: "1. Solution exception:Index was out of range..."
This change catches the "OSM 3D" component's error message:
```1. Solution exception:Index was out of range. Must be non-negative and less than the size of the collection.
Parameter name: index```

The error happens when the number of _keys is not the same as the number of values in _values branches (except branches with 0 values).